requirements.txt
streamlit
pillow
pyyaml
dash
dash-bootstrap-components
globox